rename generate-blog-components in-repo-addon to just generate-blog
