+++ ember-cli-build.js
@@ -13,7 +13,7 @@ const typescript = require('broccoli-typescript-compiler').default;
 const glob = require('glob');
 const commonjs = require('rollup-plugin-commonjs');
 const resolve = require('rollup-plugin-node-resolve');
-const collectPosts = require('./lib/generate-blog-components/lib/collect-posts');
+const collectPosts = require('./lib/generate-blog/lib/collect-posts');
 const _ = require('lodash');
 
 function findAllComponents() {
