+++ lib/generate-blog/index.js
@@ -26,7 +26,7 @@ Renderer.code = function(code, language) {
 };
 
 module.exports = {
-  name: 'generate-blog-components',
+  name: 'generate-blog',
 
   preprocessTree(type, tree) {
     if (type === 'src') {
