+++ lib/generate-blog/lib/collect-posts.js
