+++ lib/generate-blog/lib/files/author/stylesheet.hbs
