+++ lib/generate-blog/lib/files/author/template.hbs
