+++ lib/generate-blog/lib/files/post/stylesheet.hbs
