+++ lib/generate-blog/lib/files/post/template.hbs
