+++ lib/generate-blog/lib/files/recent-posts/stylesheet.hbs
