+++ lib/generate-blog/lib/files/recent-posts/template.hbs
