+++ lib/generate-blog/lib/files/start-page/stylesheet.hbs
