+++ lib/generate-blog/lib/files/start-page/template.hbs
