+++ lib/generate-blog/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "generate-blog-components",
+  "name": "generate-blog",
   "keywords": [
     "ember-addon"
   ]
